refactor: move CreateContentMenu into sidebar dir
